refactor(cli): extract terminal interaction
