+++ nanobot/cli/commands.py
@@ -4,15 +4,14 @@
 
 import asyncio
 import os
-import select
 import signal
 import sys
 import time
 from collections.abc import Awaitable, Callable, Coroutine, Iterable
-from contextlib import nullcontext, suppress
+from contextlib import suppress
 from pathlib import Path
 from types import FrameType
-from typing import TYPE_CHECKING, Any, Literal, cast
+from typing import TYPE_CHECKING, Any, cast
 
 if TYPE_CHECKING:
     from nanobot.gateway.runtime import GatewayRuntime
@@ -56,17 +55,8 @@ def _set_nanobot_logs(enabled: bool) -> None:
         logger.disable("nanobot")
 
 
-from prompt_toolkit import PromptSession, print_formatted_text  # noqa: E402
-from prompt_toolkit.application import run_in_terminal  # noqa: E402
-from prompt_toolkit.formatted_text import ANSI, HTML  # noqa: E402
-from prompt_toolkit.history import FileHistory  # noqa: E402
-from prompt_toolkit.key_binding import KeyBindings  # noqa: E402
-from prompt_toolkit.key_binding.key_processor import KeyPressEvent  # noqa: E402
-from prompt_toolkit.keys import Keys  # noqa: E402
-from prompt_toolkit.patch_stdout import patch_stdout  # noqa: E402
 from pydantic import ValidationError  # noqa: E402
 from rich.console import Console  # noqa: E402
-from rich.markdown import Markdown  # noqa: E402
 from rich.markup import escape  # noqa: E402
 from rich.table import Table  # noqa: E402
 from rich.text import Text  # noqa: E402
@@ -76,13 +66,12 @@ from nanobot import optional_features as feature_support  # noqa: E402
 from nanobot.agent.hooks import create_file_edit_activity_hook  # noqa: E402
 from nanobot.agent.loop import AgentLoop  # noqa: E402
 from nanobot.bus.outbound_events import (  # noqa: E402
-    ProgressEvent,
-    RetryWaitEvent,
     StreamDeltaEvent,
     StreamedResponseEvent,
     StreamEndEvent,
     outbound_event_from_message,
 )
+from nanobot.cli import terminal as cli_terminal  # noqa: E402
 from nanobot.cli.gateway import create_gateway_app  # noqa: E402
 from nanobot.cli.provider import provider_app  # noqa: E402
 from nanobot.cli.stream import StreamRenderer, ThinkingSpinner  # noqa: E402
@@ -112,39 +101,14 @@ from nanobot.webui.build import (  # noqa: E402
 )
 from nanobot.webui.sidebar_state import read_webui_sidebar_state  # noqa: E402
 
+# Backward-compatible import for callers that used the former module location.
+SafeFileHistory = cli_terminal.SafeFileHistory
+
 
 def _signal_name(signum: int) -> str:
     with suppress(ValueError):
         return signal.Signals(signum).name
     return f"signal {signum}"
-
-
-def _ensure_interactive_tty_mode() -> None:
-    """Restore interactive line input after a raw-mode TTY leak."""
-    try:
-        fd = sys.stdin.fileno()
-        if not os.isatty(fd):
-            return
-    except Exception:
-        return
-
-    with suppress(Exception):
-        import termios
-
-        attrs = termios.tcgetattr(fd)
-        required_lflag = termios.ISIG | termios.ICANON | termios.ECHO
-        blocked_input_flags = getattr(termios, "IGNCR", 0) | getattr(termios, "INLCR", 0)
-        if (
-            (attrs[3] & required_lflag) == required_lflag
-            and attrs[0] & termios.ICRNL
-            and not attrs[0] & blocked_input_flags
-        ):
-            return
-        attrs[0] = (attrs[0] | termios.ICRNL) & ~blocked_input_flags
-        attrs[3] |= required_lflag
-        termios.tcsetattr(fd, termios.TCSANOW, attrs)
-        termios.tcflush(fd, termios.TCIFLUSH)
-        logger.debug("Restored foreground gateway TTY mode")
 
 
 def _install_gateway_shutdown_handlers(
@@ -217,18 +181,6 @@ def _commit_dream_changes(memory: Any) -> str | None:
     return memory.git.auto_commit(message)
 
 
-class SafeFileHistory(FileHistory):
-    """FileHistory subclass that sanitizes surrogate characters on write.
-
-    On Windows, special Unicode input (emoji, mixed-script) can produce
-    surrogate characters that crash prompt_toolkit's file write.
-    See issue #2846.
-    """
-
-    def store_string(self, string: str) -> None:
-        super().store_string(_sanitize_surrogates(string))
-
-
 app = typer.Typer(
     name="nanobot",
     context_settings={"help_option_names": ["-h", "--help"]},
@@ -237,9 +189,6 @@ app = typer.Typer(
 )
 
 console = Console()
-EXIT_COMMANDS = {"exit", "quit", "/exit", "/quit", ":q"}
-_REASONING_SENTENCE_ENDINGS = (".", "!", "?", "。", "！", "？")
-_REASONING_FLUSH_CHARS = 60
 
 _HEARTBEAT_PREAMBLE = (
     "[Your response will be delivered directly to the user's messaging app. "
@@ -303,335 +252,6 @@ def _pick_heartbeat_target_from_sessions(
         if channel in enabled and chat_id:
             return channel, chat_id
     return "cli", "direct"
-
-
-# ---------------------------------------------------------------------------
-# CLI input: prompt_toolkit for editing, paste, history, and display
-# ---------------------------------------------------------------------------
-
-_PROMPT_SESSION: PromptSession[str] | None = None
-_saved_term_attrs: list[Any] | None = None  # original termios settings, restored on exit
-
-
-def _flush_pending_tty_input() -> None:
-    """Drop unread keypresses typed while the model was generating output."""
-    try:
-        fd = sys.stdin.fileno()
-        if not os.isatty(fd):
-            return
-    except Exception:
-        return
-
-    with suppress(Exception):
-        import termios
-
-        termios.tcflush(fd, termios.TCIFLUSH)
-        return
-
-    with suppress(Exception):
-        while True:
-            ready, _, _ = select.select([fd], [], [], 0)
-            if not ready:
-                break
-            if not os.read(fd, 4096):
-                break
-
-
-def _restore_terminal() -> None:
-    """Restore terminal to its original state (echo, line buffering, etc.)."""
-    if _saved_term_attrs is None:
-        return
-    with suppress(Exception):
-        import termios
-
-        termios.tcsetattr(sys.stdin.fileno(), termios.TCSADRAIN, _saved_term_attrs)
-
-
-def _build_cli_key_bindings() -> KeyBindings:
-    """Key bindings for the interactive prompt.
-
-    Behaviour:
-      * Enter       -> submit the current input (keeps the familiar
-                       single-line Enter-to-send feel even though the buffer
-                       is multiline-capable).
-      * Alt+Enter   -> insert a newline for multi-line input.
-      * Shift+Enter -> insert a newline on terminals that emit the CSI-u
-                       (kitty / fixterms) keyboard-protocol encoding for it.
-    """
-    # prompt_toolkit does not recognize CSI-u, so register its Shift+Enter
-    # sequence as a best-effort addition without overriding existing mappings.
-    with suppress(Exception):
-        from prompt_toolkit.input import ansi_escape_sequences as _aes
-
-        _aes.ANSI_SEQUENCES.setdefault("\x1b[13;2u", Keys.ControlF3)
-
-    kb = KeyBindings()
-
-    @kb.add("enter")
-    def _(event: KeyPressEvent) -> None:
-        event.current_buffer.validate_and_handle()
-
-    @kb.add("escape", "enter")  # Alt+Enter / Meta+Enter (ESC + CR, "\x1b\r")
-    def _(event: KeyPressEvent) -> None:
-        event.current_buffer.insert_text("\n")
-
-    # LF-as-Enter terminals send Alt+Enter as ESC + LF rather than ESC + CR.
-    @kb.add("escape", Keys.ControlJ)  # Alt+Enter on LF-as-Enter terminals
-    def _(event: KeyPressEvent) -> None:
-        event.current_buffer.insert_text("\n")
-
-    @kb.add(Keys.ControlF3)  # Shift+Enter on CSI-u capable terminals
-    def _(event: KeyPressEvent) -> None:
-        event.current_buffer.insert_text("\n")
-
-    return kb
-
-
-def _init_prompt_session() -> None:
-    """Create the prompt_toolkit session with persistent file history."""
-    global _PROMPT_SESSION, _saved_term_attrs
-
-    # Save terminal state so we can restore it on exit
-    with suppress(Exception):
-        import termios
-
-        _saved_term_attrs = termios.tcgetattr(sys.stdin.fileno())
-
-    from nanobot.config.paths import get_cli_history_path
-
-    history_file = get_cli_history_path()
-    history_file.parent.mkdir(parents=True, exist_ok=True)
-
-    _PROMPT_SESSION = PromptSession(
-        history=SafeFileHistory(str(history_file)),
-        enable_open_in_editor=False,
-        # Multiline-capable buffer; Enter still submits via the custom key
-        # bindings, while Alt+Enter adds a newline.
-        multiline=True,
-        key_bindings=_build_cli_key_bindings(),
-    )
-
-
-def _make_console() -> Console:
-    return Console(file=sys.stdout)
-
-
-def _render_interactive_ansi(render_fn: Callable[[Console], None]) -> str:
-    """Render Rich output to ANSI so prompt_toolkit can print it safely."""
-    ansi_console = Console(
-        force_terminal=sys.stdout.isatty(),
-        color_system=cast(
-            Literal["auto", "standard", "256", "truecolor", "windows"],
-            console.color_system or "standard",
-        ),
-        width=console.width,
-    )
-    with ansi_console.capture() as capture:
-        render_fn(ansi_console)
-    return capture.get()
-
-
-def _print_agent_response(
-    response: str,
-    render_markdown: bool,
-    metadata: dict[str, Any] | None = None,
-    show_header: bool = True,
-) -> None:
-    """Render assistant response with consistent terminal styling."""
-    console = _make_console()
-    content = response or ""
-    body = _response_renderable(content, render_markdown, metadata)
-    if show_header:
-        console.print()
-        console.print(f"[cyan]{__logo__} nanobot[/cyan]")
-    console.print(body)
-    console.print()
-
-
-def _response_renderable(
-    content: str, render_markdown: bool, metadata: dict[str, Any] | None = None
-) -> Text | Markdown:
-    """Render plain-text command output without markdown collapsing newlines."""
-    if not render_markdown:
-        return Text(content)
-    if (metadata or {}).get("render_as") == "text":
-        return Text(content)
-    return Markdown(content)
-
-
-async def _print_interactive_line(text: str) -> None:
-    """Print async interactive updates with prompt_toolkit-safe Rich styling."""
-    def _write() -> None:
-        ansi = _render_interactive_ansi(
-            lambda c: c.print(f"  [dim]↳ {text}[/dim]")
-        )
-        print_formatted_text(ANSI(ansi), end="")
-
-    await run_in_terminal(_write)
-
-
-async def _print_interactive_response(
-    response: str,
-    render_markdown: bool,
-    metadata: dict[str, Any] | None = None,
-) -> None:
-    """Print async interactive replies with prompt_toolkit-safe Rich styling."""
-    def _write() -> None:
-        content = response or ""
-
-        def _render(target: Console) -> None:
-            target.print()
-            target.print(f"[cyan]{__logo__} nanobot[/cyan]")
-            target.print(_response_renderable(content, render_markdown, metadata))
-            target.print()
-
-        ansi = _render_interactive_ansi(_render)
-        print_formatted_text(ANSI(ansi), end="")
-
-    await run_in_terminal(_write)
-
-
-def _print_cli_progress_line(text: str, thinking: ThinkingSpinner | None, renderer: StreamRenderer | None = None) -> None:
-    """Print a CLI progress line, pausing the spinner if needed."""
-    if not text.strip():
-        return
-    target = renderer.console if renderer else console
-    pause = renderer.pause_spinner() if renderer else (thinking.pause() if thinking else nullcontext())
-    with pause:
-        if renderer:
-            renderer.ensure_header()
-        target.print(f"  [dim]↳ {text}[/dim]")
-
-
-class _ReasoningBuffer:
-    def __init__(self) -> None:
-        self._text = ""
-
-    def add(self, text: str) -> str | None:
-        if not text:
-            return None
-        self._text += text
-        if self._should_flush(text):
-            return self.flush()
-        return None
-
-    def flush(self) -> str | None:
-        text = self._text.strip()
-        self._text = ""
-        return text or None
-
-    def clear(self) -> None:
-        self._text = ""
-
-    def _should_flush(self, text: str) -> bool:
-        stripped = text.rstrip()
-        return (
-            "\n" in text
-            or stripped.endswith(_REASONING_SENTENCE_ENDINGS)
-            or len(self._text) >= _REASONING_FLUSH_CHARS
-        )
-
-
-def _print_cli_reasoning(text: str, thinking: ThinkingSpinner | None, renderer: StreamRenderer | None = None) -> None:
-    """Print reasoning/thinking content in a distinct style."""
-    if not text.strip():
-        return
-    target = renderer.console if renderer else console
-    pause = renderer.pause_spinner() if renderer else (thinking.pause() if thinking else nullcontext())
-    with pause:
-        if renderer:
-            renderer.ensure_header()
-        target.print(f"[dim italic]✻ {text}[/dim italic]")
-
-
-def _flush_cli_reasoning(
-    reasoning_buffer: _ReasoningBuffer,
-    thinking: ThinkingSpinner | None,
-    renderer: StreamRenderer | None = None,
-) -> None:
-    text = reasoning_buffer.flush()
-    if text:
-        _print_cli_reasoning(text, thinking, renderer)
-
-
-async def _print_interactive_progress_line(text: str, thinking: ThinkingSpinner | None, renderer: StreamRenderer | None = None) -> None:
-    """Print an interactive progress line, pausing the spinner if needed."""
-    if not text.strip():
-        return
-    if renderer:
-        with renderer.pause_spinner():
-            renderer.ensure_header()
-            renderer.console.print(f"  [dim]↳ {text}[/dim]")
-    else:
-        with thinking.pause() if thinking else nullcontext():
-            await _print_interactive_line(text)
-
-
-async def _maybe_print_interactive_progress(
-    msg: Any,
-    thinking: ThinkingSpinner | None,
-    channels_config: Any,
-    renderer: StreamRenderer | None = None,
-    reasoning_buffer: _ReasoningBuffer | None = None,
-) -> bool:
-    event = outbound_event_from_message(msg)
-    if isinstance(event, RetryWaitEvent):
-        await _print_interactive_progress_line(msg.content, thinking, renderer)
-        return True
-
-    if not isinstance(event, ProgressEvent):
-        return False
-
-    reasoning_buffer = reasoning_buffer or _ReasoningBuffer()
-
-    if event.reasoning_end:
-        if channels_config and not channels_config.show_reasoning:
-            reasoning_buffer.clear()
-        else:
-            _flush_cli_reasoning(reasoning_buffer, thinking, renderer)
-        return True
-
-    is_tool_hint = event.tool_hint
-    is_reasoning = event.reasoning or event.reasoning_delta
-    if is_reasoning:
-        if channels_config and not channels_config.show_reasoning:
-            reasoning_buffer.clear()
-            return True
-        text = reasoning_buffer.add(msg.content)
-        if text:
-            _print_cli_reasoning(text, thinking, renderer)
-        return True
-    if channels_config and is_tool_hint and not channels_config.send_tool_hints:
-        return True
-    if channels_config and not is_tool_hint and not channels_config.send_progress:
-        return True
-
-    await _print_interactive_progress_line(msg.content, thinking, renderer)
-    return True
-
-
-def _is_exit_command(command: str) -> bool:
-    """Return True when input should end interactive chat."""
-    return command.lower() in EXIT_COMMANDS
-
-
-async def _read_interactive_input_async() -> str:
-    """Read user input using prompt_toolkit (handles paste, history, display).
-
-    prompt_toolkit natively handles:
-    - Multiline paste (bracketed paste mode)
-    - History navigation (up/down arrows)
-    - Clean display (no ghost characters or artifacts)
-    """
-    if _PROMPT_SESSION is None:
-        raise RuntimeError("Call _init_prompt_session() first")
-    try:
-        with patch_stdout():
-            return await _PROMPT_SESSION.prompt_async(
-                HTML("<b fg='ansiblue'>You:</b> "),
-            )
-    except EOFError as exc:
-        raise KeyboardInterrupt from exc
 
 
 def version_callback(value: bool):
@@ -1578,7 +1198,7 @@ def webui(
     from nanobot.config.loader import resolve_config_env_vars, save_config
     from nanobot.gateway import GatewayRuntime, GatewayRuntimePaths, GatewayStartOptions
 
-    _ensure_interactive_tty_mode()
+    cli_terminal._ensure_interactive_tty_mode()
     config_path = _resolve_webui_config_path(config)
     created_config = not config_path.exists()
     if created_config:
@@ -2310,7 +1930,7 @@ def _run_gateway(
         runtime_tasks: asyncio.Future[list[Any]] | None = None
         runtime_tasks_drained = False
         shutdown_event = asyncio.Event()
-        _ensure_interactive_tty_mode()
+        cli_terminal._ensure_interactive_tty_mode()
         restore_shutdown_handlers = _install_gateway_shutdown_handlers(
             asyncio.get_running_loop(),
             shutdown_event,
@@ -2472,7 +2092,7 @@ def agent(
         raise typer.Exit(1) from exc
     restart_notice = consume_restart_notice_from_env()
     if restart_notice and should_show_cli_restart_notice(restart_notice, session_id):
-        _print_agent_response(
+        cli_terminal._print_agent_response(
             format_restart_completed_message(restart_notice.started_at_raw),
             render_markdown=False,
         )
@@ -2483,7 +2103,7 @@ def agent(
     def _make_progress(
         renderer: StreamRenderer | None = None,
     ) -> Callable[..., Awaitable[None]]:
-        reasoning_buffer = _ReasoningBuffer()
+        reasoning_buffer = cli_terminal._ReasoningBuffer()
 
         async def _cli_progress(content: str, *, tool_hint: bool = False, reasoning: bool = False, **_kwargs: Any) -> None:
             ch = agent_loop.channels_config
@@ -2492,7 +2112,7 @@ def agent(
                 if ch and not ch.show_reasoning:
                     reasoning_buffer.clear()
                 else:
-                    _flush_cli_reasoning(reasoning_buffer, _thinking, renderer)
+                    cli_terminal._flush_cli_reasoning(reasoning_buffer, _thinking, renderer)
                 return
 
             if reasoning:
@@ -2501,13 +2121,13 @@ def agent(
                     return
                 text = reasoning_buffer.add(content)
                 if text:
-                    _print_cli_reasoning(text, _thinking, renderer)
+                    cli_terminal._print_cli_reasoning(text, _thinking, renderer)
                 return
             if ch and tool_hint and not ch.send_tool_hints:
                 return
             if ch and not tool_hint and not ch.send_progress:
                 return
-            _print_cli_progress_line(content, _thinking, renderer)
+            cli_terminal._print_cli_progress_line(content, _thinking, renderer)
         return _cli_progress
 
     if message:
@@ -2529,7 +2149,7 @@ def agent(
                 print_kwargs: dict[str, Any] = {}
                 if renderer.header_printed:
                     print_kwargs["show_header"] = False
-                _print_agent_response(
+                cli_terminal._print_agent_response(
                     response.content if response else "",
                     render_markdown=markdown,
                     metadata=response.metadata if response else None,
@@ -2541,7 +2161,7 @@ def agent(
     else:
         # Interactive mode — route through bus like other channels
         from nanobot.bus.events import InboundMessage
-        _init_prompt_session()
+        cli_terminal._init_prompt_session()
         _model, _preset_tag = _model_display(runtime_config)
         _icon = runtime_config.agents.defaults.bot_icon or __logo__
         console.print(f"{_icon} Interactive mode [bold blue]({_model})[/bold blue]{_preset_tag} — type [bold]exit[/bold] or [bold]Ctrl+C[/bold] to quit\n")
@@ -2553,7 +2173,7 @@ def agent(
 
         def _handle_signal(signum: int, _frame: FrameType | None) -> None:
             sig_name = signal.Signals(signum).name
-            _restore_terminal()
+            cli_terminal._restore_terminal()
             console.print(f"\nReceived {sig_name}, goodbye!")
             sys.exit(0)
 
@@ -2573,7 +2193,7 @@ def agent(
             turn_done.set()
             turn_response: list[Any] = []
             renderer: StreamRenderer | None = None
-            reasoning_buffer = _ReasoningBuffer()
+            reasoning_buffer = cli_terminal._ReasoningBuffer()
 
             async def _consume_outbound() -> None:
                 while True:
@@ -2597,7 +2217,7 @@ def agent(
                                 print_kwargs: dict[str, Any] = {}
                                 if renderer.header_printed:
                                     print_kwargs["show_header"] = False
-                                _print_agent_response(
+                                cli_terminal._print_agent_response(
                                     msg.content,
                                     render_markdown=markdown,
                                     metadata=msg.metadata,
@@ -2606,7 +2226,7 @@ def agent(
                             turn_done.set()
                             continue
 
-                        if await _maybe_print_interactive_progress(
+                        if await cli_terminal._maybe_print_interactive_progress(
                             msg,
                             None,
                             agent_loop.channels_config,
@@ -2620,7 +2240,7 @@ def agent(
                                 turn_response.append(msg)
                             turn_done.set()
                         elif msg.content:
-                            await _print_interactive_response(
+                            await cli_terminal._print_interactive_response(
                                 msg.content,
                                 render_markdown=markdown,
                                 metadata=msg.metadata,
@@ -2636,17 +2256,19 @@ def agent(
             try:
                 while True:
                     try:
-                        _flush_pending_tty_input()
+                        cli_terminal._flush_pending_tty_input()
                         # Stop spinner before user input to avoid prompt_toolkit conflicts
                         if renderer:
                             renderer.stop_for_input()
-                        user_input = _sanitize_surrogates(await _read_interactive_input_async())
+                        user_input = _sanitize_surrogates(
+                            await cli_terminal._read_interactive_input_async()
+                        )
                         command = user_input.strip()
                         if not command:
                             continue
 
-                        if _is_exit_command(command):
-                            _restore_terminal()
+                        if cli_terminal._is_exit_command(command):
+                            cli_terminal._restore_terminal()
                             console.print("\nGoodbye!")
                             break
 
@@ -2679,7 +2301,7 @@ def agent(
                                 print_kwargs: dict[str, Any] = {}
                                 if renderer and renderer.header_printed:
                                     print_kwargs["show_header"] = False
-                                _print_agent_response(
+                                cli_terminal._print_agent_response(
                                     content,
                                     render_markdown=markdown,
                                     metadata=meta,
@@ -2688,11 +2310,11 @@ def agent(
                         elif renderer and not renderer.streamed:
                             await renderer.close()
                     except KeyboardInterrupt:
-                        _restore_terminal()
+                        cli_terminal._restore_terminal()
                         console.print("\nGoodbye!")
                         break
                     except EOFError:
-                        _restore_terminal()
+                        cli_terminal._restore_terminal()
                         console.print("\nGoodbye!")
                         break
             finally:
